fix sizing issue by setting reactive layout
