+++ src/textual/widgets/_footer.py
@@ -69,7 +69,7 @@ class FooterKey(Widget):
 
     upper_case_keys = reactive(False)
     ctrl_to_caret = reactive(True)
-    compact = reactive(True)
+    compact = reactive(True, layout=True)
 
     def __init__(
         self,
